+++ packages/react/src/index.ts
@@ -1,4 +1,4 @@
-export { RichTextEditor } from './components/RichTextEditor';
+export { RichTextEditor as EditoraEditor, RichTextEditor } from './components/RichTextEditor';
 export { Toolbar } from './components/Toolbar';
 export { EditorContent } from './components/EditorContent';
 export { InlineMenu } from './components/InlineMenu';